backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pandas
openpyxl
httpx
pytest
hypothesis
python-dotenv
